Reference

Standard DD

Binary DD

ElementaryFluxModes.DDBinary — Method

DDBinary(N, K) -> Any

Implement the Double Description method in binary form. The input variables are: -N: the stoichiometric matrix with only forward reactions. If any fluxes are fixed then N has the form [N1 w] where N1 is the stoichiometric matrix for non-fixed fluxes, and w is the columns of fixed fluxes multiplied by their flux values.

-K: initial nullspace of N, in the form [I;K*] Output: -R: binary elementary flux modes

ElementaryFluxModes.check_adjacency — Method

check_adjacency(i, j, R) -> Bool

Check adjacency of columns i and j in r by making sure that there exists no other extreme ray in R whose zero set is a superset of the intersection of the zero sets of ray i and ray j.

ElementaryFluxModes.get_efms — Method

get_efms(N::Matrix{Float64}; tol) -> Any

Calculate elementary flux modes of a stoichiometric matrix. Input N must be the stoichiometric matrix of a network with only forward reactions. Output: vector of size (n,k) of the fluxes through the n reactions in the k EFMs.

ElementaryFluxModes.make_bitmap — Method

make_bitmap(row) -> Vector{Bool}

Return a bitmap of the input vector, where entries are 1 if the vector entry is greater than zero, and zero if the entries are zero. Return an error if any entries of the vector are less than zero.

ElementaryFluxModes.rational_nullspace — Method

rational_nullspace(
    A::Matrix;
    tol
) -> Tuple{Matrix{Float64}, Any}

Helper function to calculate a nullspace of the matrix A, with all rational entries.

ElementaryFluxModes.zero_set — Method

zero_set(vec) -> Vector

Return the zero set of a vector

Utilities

ElementaryFluxModes.fix_fluxes — Method

fix_fluxes(
    S::Matrix{Float64},
    fixed_fluxes::Vector{Int64},
    flux_values::Vector{Float64}
) -> Matrix{Float64}

Function to make a convex polyhedron out of a problem with fixed fluxes. Input: S: irreversible stoichiometric matrix, aka any reversible reactions have been split into two irreversible reactions fixedfluxes: list of indices of the fixed fluxes fluxvalues: list of fixed flux values in same order as fixedfluxes Output: Sconvex: stoichiometric matrix with which to implement the DD algorithm Note: results of DDStandard need to be transformed to take into account these fixed fluxes, using cleanDDresult

ElementaryFluxModes.remove_linearly_dep_rows — Method

remove_linearly_dep_rows(
    A::Matrix{Float64};
    tol
) -> Tuple{Matrix{Float64}, Vector{Int64}}

Helper function to remove linearly dependent rows of the matrix A.

ElementaryFluxModes.reorder_ns — Method

reorder_ns(A::Matrix) -> Tuple{Matrix, Vector{Int64}}

Helper function to reorder the rows of the nullspace so that it is in the form [I; K].

ElementaryFluxModes.reversible_EFMs — Method

reversible_EFMs(
    E::Matrix{Float64},
    reversible::Vector{Int64}
) -> Matrix{Float64}

Return the EFMs in terms of the original stoichiometric matrix, with reversible reactions.